% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpa_color.R
\name{mpa_color}
\alias{mpa_color}
\title{Define MPA colors}
\usage{
mpa_color(...)
}
\arguments{
\item{...}{Character strings specifying the color names.
If empty, all colors are returned.}
}
\value{
A named vector of hex color codes.
}
\description{
This function returns the hex color codes for specified colors used by the MPA.
If no colors are specified, it returns all available colors.
}
\examples{
mpa_color("mpa_blue")
mpa_color("mpa_blue", "red")
mpa_color() # returns all colors
}
