% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpa_palette.R
\name{mpa_palette}
\alias{mpa_palette}
\title{Retrieve a Predefined Color Palette}
\usage{
mpa_palette(palette = "mpa_blues", ...)
}
\arguments{
\item{palette}{A character string specifying the name of the palette to retrieve. Accepted values are 'mpa_blues', 'complementary', 'complementary_60', 'complementary_20', 'complementary_all', and 'highlight'. Defaults to 'mpa_blues'.}

\item{...}{Additional arguments passed to the color selection function (if applicable).}
}
\value{
A list of color values corresponding to the chosen palette.
}
\description{
This function provides access to a set of predefined color palettes. It allows users to select a specific palette, each composed of various color shades. Additional parameters can modify the palette selection behavior.
}
\examples{
mpa_palette("mpa_blues")
mpa_palette("complementary")
mpa_palette("highlight")
}
