% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPAcolors-package.R
\docType{package}
\name{MPAcolors-package}
\alias{MPAcolors}
\alias{MPAcolors-package}
\title{MPAcolors: Provides color palettes used by the Swedish Medical Products Agency (MPA)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/vrognas/MPAcolors}
  \item Report bugs at \url{https://github.com/vrognas/MPAcolors/issues}
}

}
\author{
\strong{Maintainer}: Viktor Rognås \email{viktor.rognas@gmail.com} (\href{https://orcid.org/0000-0002-7837-2467}{ORCID})

}
\keyword{internal}
