% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_gen.R
\name{palette_gen}
\alias{palette_gen}
\title{Generate a Color Palette Function for Discrete Scales}
\usage{
palette_gen(palette = "complementary", direction = 1)
}
\arguments{
\item{palette}{A character string indicating the name of the color palette.
Accepted values include 'complementary' and others as defined in \code{mpa_palette}.
Defaults to 'complementary'.}

\item{direction}{An integer indicating the direction of the color scale.
Use 1 for normal order and -1 to reverse the order. Defaults to 1.}
}
\value{
Returns a function that, when called with an integer \code{n},
generates a vector of \code{n} color values from the specified palette.
This function warns if \code{n} exceeds the number of colors available in the palette.
}
\description{
\code{palette_gen} is a factory function that creates a function for generating color palettes.
This inner function is designed to be used with \code{scale_fill} and \code{scale_color} functions
in plotting contexts, providing a specific number of colors from the selected palette.
}
