% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_colour_mpa.R
\name{scale_colour_mpa_c}
\alias{scale_colour_mpa_c}
\alias{scale_color_mpa_c}
\title{Set Continuous Color Scale for ggplot2 Plots}
\usage{
scale_colour_mpa_c(palette = "highlight", direction = 1, ...)

scale_color_mpa_c(palette = "highlight", direction = 1, ...)
}
\arguments{
\item{palette}{A string specifying the palette name. Defaults to 'complementary'.}

\item{direction}{A numeric value specifying the direction of the scale. Defaults to 1.}

\item{...}{Additional arguments passed to ggplot2's discrete_scale function.}
}
\value{
A ggplot2 scale object.
}
\description{
Applies a continuous color gradient scale for color aesthetics in ggplot2 plots, using predefined color palettes.
}
