% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_mpa.R
\name{scale_fill_mpa}
\alias{scale_fill_mpa}
\title{Set Fill Scale for ggplot2 Plots Using Predefined Palettes}
\usage{
scale_fill_mpa(palette = "complementary", direction = 1, ...)
}
\arguments{
\item{palette}{A string specifying the palette name. Defaults to 'complementary'.}

\item{direction}{A numeric value specifying the direction of the scale. Defaults to 1.}

\item{...}{Additional arguments passed to ggplot2's discrete_scale function.}
}
\value{
A ggplot2 scale object.
}
\description{
Applies a discrete color scale for fill aesthetics in ggplot2 plots, using predefined color palettes.
}
